Update README for v2.0.0
The README file for the Katalon Studio Sealight Plugin has been updated to provide clearer instructions on how to use the plugin. Changes include a more detailed explanation of how to install and configure the plugin, as well as an update on creating and executing test suites. The section on contributing and companion products has been removed.